WAS Zero from Web and AGOpen
if you press "WAS zero" or move WAS slider in the config of AgOpenGPS, then this will be used to trigger the "Was set to zero function" (same function as pressing zero in the Web-config). The actual value of the slider in AgOpenGps will be ignored, every change is just a trigger!